Log console error messages to database.
Add a new database table: "console_errors".
Store timestamp, script URL and the "clipboardText" of the console error.
The clipboardText contains all the details such as error type, stack
trace, script location etc.

One major side effect of this change is we have to keep the developer
tools open to access the devtools API. We open the panel in a
detached window to minimize the fingerprintability of this setup.