Use substring search to detect leaks, add ability to search resp. leaks.
Search leaks in all requests and responses, not only third parties.
Search upper case versions of the search strings.
Add substring search for cookies, url location and referrer.

Handle cases where we cannot create an output dir.

Reduce encoding_layers and hash_layers from 3 to 2.
Add data path of the cluster.
Load all rows into memory when we no of CPU > 16. This is for
running the code on PU clusters.